GPU Element wise binary comparison

+++ Assets/LPE/DumbML/Operations/Tensor Buffers/CPU/IntCPUTensorBuffer.cs
@@ -1,6 +1,6 @@
 ﻿namespace DumbML {
     public class IntCPUTensorBuffer : CPUTensorBuffer<int> {
-        public override DType dtype => DType.Float;
+        public override DType dtype => DType.Int;
 
         public IntCPUTensorBuffer(params int[] shape) : base(shape) { }
     }
